Removed defer
Feel free to close this if you don't like the change. I think it very lightly increases performance based on a discussion from [here](https://github.com/gorilla/context/pull/12). I think defer is better suited on a function with multiple exits or coupled on an open / close situation. I love this example. I'll try and take a closer look later and maybe help with some improvements if that's welcome!